+++ mnefun/stats/__init__.py
@@ -1,0 +1,3 @@
+"""Functions for statistical analysis"""
+
+from mnefun import _paths
